requirements.txt
anyio==3.5.0
asgiref==3.5.0
asn1crypto==1.5.1
bcrypt==3.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==37.0.2
cycler==0.11.0
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.0
fastapi==0.75.0
fonttools==4.31.1
google-api-core==2.7.2
google-auth==2.6.5
google-cloud-appengine-logging==1.1.1
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.0
google-cloud-logging==3.0.0
google-cloud-secret-manager==2.10.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpc-google-iam-v1==0.12.4
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
imageio==2.16.1
Jinja2==3.1.1
kiwisolver==1.4.0
MarkupSafe==2.1.1
matplotlib==3.5.1
networkx==2.7.1
numpy==1.22.3
opencv-contrib-python==4.5.5.64
packaging==21.3
passlib==1.7.4
pg8000==1.26.1
phonenumbers==8.12.50
Pillow==9.0.1
proto-plus==1.20.3
protobuf==3.20.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
pyserial==3.5
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyWavelets==1.3.0
requests==2.27.1
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
scramp==1.4.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
tifffile==2022.3.16
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
